shopee/requirements.txt
playwright
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openpyxl
pandas
